requirements.txt
google-generativeai>=0.5.2
PyPDF2